eaigaq_project/requirements.txt
asgiref==3.8.1
requests
Django==5.1.2
gunicorn
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
face_recognition
celery
gevent
eventlet
django-filter
channels
aiortc
redis
numpy
aioredis
cython